nested scss not showing, other changes made

+++ src/components/IndividualProject.js
@@ -1,12 +1,12 @@
-import React, {useState} from 'react';
-import {FaTrashAlt} from 'react-icons/fa';
-import {useProjectsValue, useSelectedProjectValue} from '../context';
-import {firebase} from '../firebase'
+import React, { useState } from 'react';
+import { FaTrashAlt } from 'react-icons/fa';
+import { useProjectsValue, useSelectedProjectValue } from '../context';
+import { firebase } from '../firebase'
 
-const IndividualProject = ({project}) => {
+const IndividualProject = ({ project }) => {
     const [showConfirm, setShowConfirm] = useState(false);
-    const {projects, setProjects} = useProjectsValue()
-    const {setSelectedProject} = useSelectedProjectValue()
+    const { projects, setProjects } = useProjectsValue()
+    const { setSelectedProject } = useSelectedProjectValue()
 
     const deleteProject = docId => {
         firebase
@@ -14,28 +14,28 @@ const IndividualProject = ({project}) => {
             .collection('projects')
             .doc(docId)
             .delete()
-            .then(()=>{
+            .then(() => {
                 setProjects([...projects])
                 setSelectedProject('INBOX')
             })
     }
 
-    return(
+    return (
         <>
-        <span className='sidebar__dot'>•</span>
-        <span className='sidebar__project-name'>{project.name}</span>
-        <span className='sidebar__project-delete' data-testid='delete-project' onClick={()=>setShowConfirm(!showConfirm)}><FaTrashAlt />
-        {showConfirm && (
-            <div className='project-delete-modal'>
-                <div className='project-delete-modal__inner'>
-                    <p>Are you sure you want to delete this project?</p>
-                    <button type='button' onClick={()=>deleteProject(project.docId)}>Delete
-                    <span onClick={() => setShowConfirm(!showConfirm)}>Cancel</span>
+            <span className='sidebar__dot'>•</span>
+            <span className='sidebar__project-name'>{project.name}</span>
+            <span className='sidebar__project-delete' data-testid='delete-project' onClick={() => setShowConfirm(!showConfirm)}><FaTrashAlt />
+                {showConfirm && (
+                    <div className='project-delete-modal'>
+                        <div className='project-delete-modal__inner'>
+                            <p>Are you sure you want to delete this project?</p>
+                            <button type='button' onClick={() => deleteProject(project.docId)}>Delete
                     </button>
-                </div>
-            </div>
-        )}
-        </span>
+                            <span onClick={() => setShowConfirm(!showConfirm)}>Cancel</span>
+                        </div>
+                    </div>
+                )}
+            </span>
         </>
     )
 }
